feature(user-profile): Add function to check wether the user logged in, refactor AuthForm CSS, simplify UserProfile page by replacing custom components with ProfileMenu and UserProfileForm, and remove unused CSS styles.
