+++ src/components/organisms/auth-form/AuthForm.css
@@ -1,11 +1,13 @@
 /** Desktop first */
+/* Register */
 .register {
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     width: 40%;
-    height: 98%;
+    height: 110%;
+    padding: 20px;
     background-color: hsl(0, 0%, 100%);
     gap: 10px;
     border: 1px;
@@ -17,19 +19,21 @@
     flex-direction: column;
     justify-content: center;
     width: 90%;
-    height: 55%;
+    height: 90%;
     gap: 10px;
 }
 
+/* Login */
 .login {
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     width: 40%;
-    height: 75%;
+    height: 85%;
+    padding: 20px;
     background-color: hsl(0, 0%, 100%);
-    gap: 30px;
+    gap: 10px;
     border: 1px;
     border-radius: 10px;
 }
@@ -39,7 +43,7 @@
     flex-direction: column;
     justify-content: center;
     width: 90%;
-    height: 70%;
+    height: 90%;
     gap: 16px;
 }
 
@@ -136,7 +140,7 @@ input:focus {
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    width: 90%;
+    width: 100%;
     height: 35%;
     gap: 10px;
 }
