+++ src/components/organisms/profile-menu/ProfileMenu.tsx
@@ -1,0 +1,29 @@
+import Image from "../../atoms/Image";
+import userIcon from "../../../assets/svg/user-icon-active.svg";
+import userClass from "../../../assets/svg/user-class-nonactive.svg";
+import userOrder from "../../../assets/svg/user-order-nonactive.svg";
+import "./ProfileMenu.css";
+
+const ProfileMenu = () => {
+  return (
+    <div style={{ display: "flex", flexDirection: "column", gap: "1.5em" }}>
+      <div style={{ display: "flex", flexDirection: "column", gap: "0.625em" }}>
+        <h2>Ubah Profil</h2>
+        <p>Ubah data diri Anda</p>
+      </div>
+      <ul className="profile-menu">
+        <li className="active">
+          <Image src={userIcon} alt="User icon" /> Profil Saya
+        </li>
+        <li>
+          <Image src={userClass} alt="User class icon" /> Kelas Saya
+        </li>
+        <li>
+          <Image src={userOrder} alt="User order icon" /> Pesanan Saya
+        </li>
+      </ul>
+    </div>
+  );
+};
+
+export default ProfileMenu;
