+++ src/components/organisms/user-profile-form/UserProfileForm.css
@@ -1,0 +1,81 @@
+.user-profile-form {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    background-color: hsl(0, 0%, 100%);
+    padding: 1.2em;
+    border: 1px solid hsla(265, 10%, 23%, 0.12);
+    border-radius: 10px;
+    gap: 1.2em;
+    font-size: 20px;
+    list-style: none;
+}
+
+.user-profile-form img {
+    width: 92px;
+    height: 92px;
+}
+
+hr {
+    border: 1px solid hsla(265, 10%, 23%, 0.12);
+}
+
+.form-group {
+    position: relative;
+    margin: 20px 0;
+}
+
+.form-control {
+    width: 90%;
+    padding: 10px;
+    font-size: 16px;
+    border: 1px solid hsl(0, 0%, 80%);
+    border-radius: 5px;
+    outline: none;
+}
+
+.form-control:focus {
+    border-color: hsl(134, 61%, 41%);
+}
+
+.form-label {
+    position: absolute;
+    top: 50%;
+    left: 10px;
+    font-size: 16px;
+    color: hsl(0, 0%, 67%);
+    background-color: hsl(0, 0%, 100%);
+    padding: 0 5px;
+    transform: translateY(-50%);
+    transition: 0.2s ease all;
+    pointer-events: none;
+}
+
+.form-control:focus+.form-label,
+.form-control:not(:placeholder-shown)+.form-label {
+    top: -10px;
+    font-size: 12px;
+    color: hsl(134, 61%, 41%);
+}
+
+.form-select {
+    width: 100%;
+    height: 57px;
+    padding: 10px;
+    font-size: 16px;
+    border: 1px solid hsl(0, 0%, 80%);
+    border-radius: 5px;
+    background-color: hsl(0, 0%, 100%);
+    outline: none;
+}
+
+.form-select:focus {
+    border-color: hsl(134, 61%, 41%);
+}
+
+.save-button {
+    position: relative;
+    width: 112px;
+    height: 46px;
+    margin-left: auto;
+}
